rag-prodiction-architecture/requirements.txt
fastapi
uvicorn
faiss-cpu
sentence-transformers
numpy
transformers
torch
